% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{affairs}
\alias{affairs}
\title{Political Affairs}
\format{
A data frame with 9146 rows and 7 variables:
\describe{
  \item{affairID}{the unique ID of the affair}
  \item{type}{the type of the affair}
  \item{date}{the date the affair was initially submitted}
  \item{session}{the parliament session number the affair was submitted}
  \item{topics}{topic areas of the affair}
  \item{text}{text of the affair}
  \item{pID}{the ID of the author (parliament member)}
  \item{party}{the party of the author}
}
}
\source{
ws-old.parliament.ch / self combined
}
\usage{
affairs
}
\description{
A dataset containing all affairs from 2015 to 2019 of the Swiss parliament.
The date was downloaded from Swiss Parliament API.
}
\keyword{datasets}
