% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_files}
\alias{download_files}
\title{Download files for excercises.}
\usage{
download_files(path = getwd())
}
\arguments{
\item{path}{The path to download the files to. Default: current working directory.}
}
\value{
Files used in Benjamin Schlegel's R Book.
}
\description{
Download files for excercises.
}
\examples{
download_files()
}
