% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partystrength}
\alias{partystrength}
\title{Party Streng Swiss Nation Council}
\format{
A data frame with 702 rows and 3 variables:
\describe{
  \item{Partei}{party}
  \item{Jahr}{year of the election}
  \item{Staerke}{strength of the party}
}
}
\source{
Bundesamt für Statistik BFS, commercial usage only with approval of them
}
\usage{
partystrength
}
\description{
A dataset containing data on the party streng on the Swiss National Council from 1991-2023
}
\keyword{datasets}
