% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world}
\alias{world}
\title{Country Data from UN}
\format{
A data frame with 2141 rows and 12 variables:
\describe{
  \item{code}{UN country code}
  \item{country}{Name of the country}
  \item{year}{Year of the value}
  \item{gdppc}{Gross domestic product (GDP) per capity}
  \item{popgrow}{Growth of the population}
  \item{fertility}{Children per women}
  \item{infantmortality}{Infant mortality for both sexes (per 1,000 live births)}
  \item{maternalmortality}{Maternal mortality ratio (deaths per 100,000 population)}
  \item{lifeexpall}{Life expectancy at birth for both sexes (years)}
  \item{lifeexpmale}{Life expectancy at birth for males (years)}
  \item{lifeexpfemale}{Life expectancy at birth for females (years)}
  \item{homicide}{Intentional homicide rates per 100,000}
}
}
\source{
undata. A world of information. \url{https://data.un.org/}
}
\usage{
world
}
\description{
A data set with country data of different years
}
\keyword{datasets}
